+++ components/Modal/CoordinationSchedule/Components/Label/Label.module.scss
