+++ components/Modal/CoordinationSchedule/Components/StepView/StepView.module.scss
